crossdock - simplify behavior dir
