+++ crossdock/behavior/echo/behavior.go
@@ -18,43 +18,47 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package main
+package echo
 
 import (
 	"github.com/yarpc/yarpc-go/crossdock-go"
-	"github.com/yarpc/yarpc-go/crossdock/behavior/echo"
-	"github.com/yarpc/yarpc-go/crossdock/behavior/errors"
-	"github.com/yarpc/yarpc-go/crossdock/behavior/gauntlet"
-	"github.com/yarpc/yarpc-go/crossdock/behavior/headers"
-	"github.com/yarpc/yarpc-go/crossdock/behavior/tchclient"
-	"github.com/yarpc/yarpc-go/crossdock/behavior/tchserver"
-	"github.com/yarpc/yarpc-go/crossdock/server"
+	"github.com/yarpc/yarpc-go/crossdock/behavior/params"
 )
 
-func main() {
-	server.Start()
-	crossdock.Start(dispatch)
+// echoEntry is an entry emitted by the echo behaviors.
+type echoEntry struct {
+	crossdock.Entry
+
+	Transport string `json:"transport"`
+	Encoding  string `json:"encoding"`
+	Server    string `json:"server"`
 }
 
-func dispatch(s crossdock.Sink, behavior string, ps crossdock.Params) {
-	switch behavior {
-	case "raw":
-		echo.Raw(s, ps)
-	case "json":
-		echo.JSON(s, ps)
-	case "thrift":
-		echo.Thrift(s, ps)
-	case "errors":
-		errors.Run(s, ps)
-	case "headers":
-		headers.Run(s, ps)
-	case "tchclient":
-		tchclient.Run(s, ps)
-	case "tchserver":
-		tchserver.Run(s, ps)
-	case "thriftgauntlet":
-		gauntlet.Run(s, ps)
-	default:
-		crossdock.Skipf(s, "unknown behavior %q", behavior)
+// echoSink wraps a sink to emit echoEntry entries instead.
+type echoSink struct {
+	crossdock.Sink
+
+	Transport string
+	Encoding  string
+	Server    string
+}
+
+func (s echoSink) Put(e interface{}) {
+	s.Sink.Put(echoEntry{
+		Entry:     e.(crossdock.Entry),
+		Transport: s.Transport,
+		Encoding:  s.Encoding,
+		Server:    s.Server,
+	})
+}
+
+// createEchoSink wraps a Sink to have transport, encoding, and server
+// information.
+func createEchoSink(encoding string, s crossdock.Sink, p crossdock.Params) crossdock.Sink {
+	return echoSink{
+		Sink:      s,
+		Transport: p.Param(params.Transport),
+		Encoding:  encoding,
+		Server:    p.Param(params.Server),
 	}
 }
