+++ crossdock/behavior/random/rand.go
